Removed `MediaSmallRow`
Moved it to `MediaSmall`.
