+++ app/src/main/java/com/imashnake/animite/features/home/MediaSmall.kt
@@ -1,12 +1,16 @@
 package com.imashnake.animite.features.home
 
 import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentHeight
+import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
@@ -22,7 +26,40 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import coil.compose.AsyncImage
+import com.imashnake.animite.MediaListQuery
+import com.imashnake.animite.type.MediaType
 import com.imashnake.animite.R as Res
+
+/**
+ * TODO: This component is hardly reusable, make changes to the API.
+ * A [LazyRow] of [MediaSmall]s.
+ *
+ * @param mediaList A list of [MediaListQuery.Medium]s.
+ * @param onItemClick Action that is propagated to the [MediaSmall]s, it takes a unique id.
+ */
+@Composable
+fun MediaSmallRow(mediaList: List<MediaListQuery.Medium?>, onItemClick: (itemId: Int?) -> Unit) {
+    LazyRow(
+        horizontalArrangement = Arrangement.spacedBy(dimensionResource(Res.dimen.small_padding)),
+        contentPadding = PaddingValues(
+            start = dimensionResource(Res.dimen.large_padding),
+            end = dimensionResource(Res.dimen.large_padding)
+        )
+    ) {
+        items(mediaList.filterNotNull()) { media ->
+            MediaSmall(
+                image = media.coverImage?.extraLarge,
+                // TODO: Do something about this chain.
+                label = media.title?.romaji ?:
+                media.title?.english ?:
+                media.title?.native ?: "",
+                onClick = {
+                    onItemClick(media.id)
+                }
+            )
+        }
+    }
+}
 
 /**
  * TODO: This component is hardly reusable, make changes to the API.
@@ -69,7 +106,7 @@ fun MediaSmall(image: String?, label: String?, onClick: () -> Unit) {
                 )
 
                 Text(
-                    text = label.orEmpty(),
+                    text = label,
                     color = MaterialTheme.colorScheme.onSurfaceVariant,
                     style = MaterialTheme.typography.labelLarge,
                     maxLines = 2,
@@ -87,6 +124,35 @@ fun MediaSmall(image: String?, label: String?, onClick: () -> Unit) {
                 )
             }
     }
+}
+
+/**
+ * Example:
+ *
+ * **Anime:** Sono Bisque Doll wa Koi wo Suru;
+ * **ID:** 132405.
+ */
+@Preview
+@Composable
+fun PreviewMediaSmallRow() {
+    MediaSmallRow(
+        mediaList = List(10) {
+            MediaListQuery.Medium(
+                id = it,
+                type = MediaType.ANIME,
+                title = MediaListQuery.Title(
+                    romaji = "Sono Bisque Doll wa Koi wo Suru",
+                    english = null,
+                    native = null
+                ),
+                coverImage = MediaListQuery.CoverImage(
+                    extraLarge = "https://s4.anilist.co/file/anilistcdn/media/anime/cover/large/bx132405-qP7FQYGmNI3d.jpg",
+                    large = null
+                )
+            )
+        },
+        onItemClick = {  }
+    )
 }
 
 @Preview
